Added js script. Modified gitignore.
